Move cli commands to the zebra-cli library
